requirements.txt
asgiref==3.8.1
Django==5.0
djangorestframework==3.15.2
pillow==10.4.0
pip==23.2.1
setuptools==65.5.0
sqlparse==0.5.1